requirements-dev.txt
pytest>=7.0
pytest-asyncio>=0.21
python-can>=4.3
